% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debuter.R
\docType{package}
\name{debuter-package}
\alias{debuter}
\alias{debuter-package}
\title{Package pour la formation "Débuter en R"
Contient les vignettes (les slides de cours), le jeu de données "fruits" et les tutoriels. https://vguillemot.github.io/debuter/}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
